Moves 'countTotalAndAverage' to the Shared folder.
