+++ fec-client/src/components/Ratings/RatingsBreakdown/ReviewBreakdown.jsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import styled from 'styled-components';
 import ReviewBar from './ReviewBar.jsx';
 import Stars from '../../Shared/Stars.jsx';
-import countTotalAndAverage from '../countTotalAndAverage.js';
+import countTotalAndAverage from '../../Shared/countTotalAndAverage.js';
 
 const ReviewBreakdownContainer = styled.div`
 display: flex;
