+++ fec-client/src/components/Shared/countTotalAndAverage.js
